Insight reputation

+++ insight-reputation/src/main/java/com/todongsan/insightreputation/reputation/entity/Reputation.java
@@ -106,7 +106,7 @@ public class Reputation extends BaseEntity {
         }
     }
 
-    public void updatePredictionAccuracy(double accuracy) {
-        this.predictionAccuracy = BigDecimal.valueOf(accuracy);
+    public void updatePredictionAccuracy(BigDecimal accuracy) {
+        this.predictionAccuracy = accuracy;
     }
 }
